results page merging
